+++ tests/test_MTPviewer2.py
@@ -29,15 +29,14 @@ from lib.rootdir import getrootdir
 from EOLpython.Qlogger.messageHandler import QLogger
 
 logger = QLogger("EOLlogger")
+# Set environment var to indicate we are in testing mode
+# Need this to logger won't try to open message boxes
+logger.setDisableMessageBox(True)
 
 
 class TESTMTPviewer(unittest.TestCase):
 
     def setUp(self):
-
-        # Set environment var to indicate we are in testing mode
-        # Need this to logger won't try to open message boxes
-        os.environ["TEST_FLAG"] = "true"
 
         # For testing, we want to capture the log messages in a buffer so we
         # can compare the log output to what we expect.