+++ tests/test_readascii_parms.py
@@ -16,7 +16,6 @@
 #
 # COPYRIGHT:   University Corporation for Atmospheric Research, 2020
 ##############################################################################
-import os
 import unittest
 from util.readascii_parms import AsciiParms
 
@@ -35,14 +34,10 @@ class TESTreadascii_parms(unittest.TestCase):
         loglevel = logging.INFO
         logger.initStream(self.stream, loglevel)
 
-        # Set environment var to indicate we are in testing mode
-        os.environ["TEST_FLAG"] = "true"
-
     def test_open(self):
         # Test that code finds ascii_parms file or exits with useful error
         self.ascii_parms = AsciiParms("nonexistentFile")
         self.assertFalse(self.ascii_parms.open())
 
     def tearDown(self):
-        if "TEST_FLAG" in os.environ:
-            del os.environ['TEST_FLAG']
+        pass